regional_hostname: move migrations test into new file
